debugging

+++ src/base/ComputeDFTMatrix.cu
@@ -12,6 +12,6 @@ __global__ void ComputeDFTMatrix(__half* dft_matrix_batch_RE,
   int collum_id = inter_matrix_id / 16;
 
   float phase = (static_cast<float>(2 * row_id * collum_id) * M_PI) / 16.0;
-  dft_matrix_batch_RE[thread_id] = __float2half(cosf(phase));
-  dft_matrix_batch_IM[thread_id] = __float2half(-sinf(phase));
+  dft_matrix_batch_RE[thread_id] = __float2half(cos(phase));
+  dft_matrix_batch_IM[thread_id] = __float2half(-sin(phase));
 }
