+++ src/base/Radix2.cu
@@ -24,8 +24,8 @@ __global__ void Radix2Kernel(__half* input_data_RE, __half* input_data_IM,
   //The twiddle factor for the first point is 1 -> only the second point has to
   //be modified
   float phase = (-M_PI * memory_point1_offset) / sub_fft_length;
-  __half twiddle_RE = __float2half(cosf(phase));
-  __half twiddle_IM = __float2half(sinf(phase));
+  __half twiddle_RE = __float2half(cos(phase));
+  __half twiddle_IM = __float2half(sin(phase));
 
   //Fetch current data once from global memory to use it twice
   __half input_RE = input_data_RE[memory_point2_offset];
